Major refactoring.
We now collect all statements and assignments in a ``{ … }`` block
before executing them. This affords just-in-time evaluation of
variables and arbitrary stateent ordering.

Also, this allows us to add a test framework that compares an
OpenSCAD-modelled object to the same thing modelled by Python code
to the same thing modelled by the "real" OpenSCAD.